精确字段类型，尽量使用Byte、Short类型 重构oauth2client 重构oauth2实现逻辑

+++ anan-cloudadviced/anan-platformserver/src/main/java/top/fosin/anan/platform/modules/dictionary/dto/DictionaryDTO.java
@@ -1,17 +1,16 @@
 package top.fosin.anan.platform.modules.dictionary.dto;
 
-                                                                                                                                                                                    
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
-import java.io.Serializable;
 import lombok.Data;
-import org.springframework.format.annotation.DateTimeFormat;
-import top.fosin.anan.core.util.DateTimeUtil;
-import java.util.Date;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
+import org.springframework.format.annotation.DateTimeFormat;
+import top.fosin.anan.core.util.DateTimeUtil;
 import top.fosin.anan.data.entity.Id;
+
+import java.util.Date;
 /**
  * 系统通用字典表(anan_dictionary)DTO
  *
@@ -28,7 +27,7 @@ public class DictionaryDTO extends Id<Long> {
     private String name;
 
     @ApiModelProperty(value = "字典类别，区别字典的大分类，取值于表anan_dictionary.code = 1数据", required = true, example = "Integer")
-    private Integer type;
+    private Byte type;
 
     @ApiModelProperty(value = "字典作用域，以字典类别为前提，在字典类别基础上再次细化分类字典", example = "String")
     private String scope;
