+++ anan-cloudadviced/anan-platformserver/src/main/java/top/fosin/anan/platform/modules/dictionary/po/DictionaryDetail.java
@@ -50,12 +50,12 @@ public class DictionaryDetail extends IdCreateUpdateDeletePO<Long> {
   @Basic
   @ApiModelProperty(value = "顺序，用于显示数据时的顺序，数值越小越靠前", required = true)
   @Column(name = "sort", nullable = false)
-  private Integer sort;
+  private Short sort;
 
   @Basic
   @ApiModelProperty(value = "使用状态：0=启用，1=禁用，具体取值于字典表anan_dictionary.id=11", required = true)
   @Column(name = "status", nullable = false)
-  private Integer status;
+  private Byte status;
 
   @Basic
   @ApiModelProperty(value = "标准代码，该字段通常用于对接标准字典")
@@ -70,7 +70,7 @@ public class DictionaryDetail extends IdCreateUpdateDeletePO<Long> {
   @Basic
   @ApiModelProperty(value = "使用标志：0=未使用，1=已使用，已使用的字典就不能再修改name属性", required = true)
   @Column(name = "used", nullable = false)
-  private Integer used = 0;
+  private Byte used = 0;
 
   @Basic
   @ApiModelProperty(value = "字典说明")
