+++ anan-cloudadviced/anan-platformserver/src/main/java/top/fosin/anan/platform/modules/dictionary/vo/DictionaryVO.java
@@ -1,13 +1,9 @@
 package top.fosin.anan.platform.modules.dictionary.vo;
 
-                                                                                                                                                                            
+
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
-import java.io.Serializable;
 import lombok.Data;
-import org.springframework.format.annotation.DateTimeFormat;
-import top.fosin.anan.core.util.DateTimeUtil;
-import java.util.Date;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
 import top.fosin.anan.data.entity.Id;
@@ -27,7 +23,7 @@ public class DictionaryVO extends Id<Long> {
     private String name;
 
     @ApiModelProperty(value = "字典类别，区别字典的大分类，取值于表anan_dictionary.code = 1数据")
-    private Integer type;
+    private Byte type;
 
     @ApiModelProperty(value = "字典作用域，以字典类别为前提，在字典类别基础上再次细化分类字典")
     private String scope;
